bug fix in publishers.py and tests.py update
Implementing a  regex pattern matching to spot the prefix in the DOIs resulted to be a suboptimal solution in the case of invalid receiving dois, which could actually be invalid because of an error in the prefix format.
In order to optimize our results, we decided to implement an if-else structure so to try to match the prefix structure in the doi before the first "/" character and maximise the chances of making a successful API request for the publisher's identification. Indeed, in this way we avoid wasting the possibility to spot the addressed publisher because of extra characters in the cited DOI.
At the same time, we addressed the possibility that the prefix itself is not compliant in the else clause: in this latter case, we just stick to the the very first implementation of the prefix extraction, which was that of considering as prefix all the characters preceding the first "/" character. However, in this case, the publisher won't be identified.

All the existing tests were updated, so to test the functions with the extra parameters added. The structures of the sample data were refactored where necessary.